pyproject.toml
[project]
name = "star-chart"
version = "0.1.0"
description = "A simple set of CLI tools designed to automatically download and parse astronomical observation data into a GeoPackage for star chart authoring."
authors = [{name = "João Victor Teófilo Salgado", email = "jvictorteo2000@hotmail.com"}]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "astroquery>=0.4.9.post1",
    "duckdb>=1.2.0",
    "geopandas>=1.0.1",
    "skyfield>=1.49",
    "typer>=0.15",
]

[project.scripts]
star-chart = "star_chart.cli:app"

[tool.uv]
package = true

[tool.setuptools]
packages = ["star_chart"]
